refactor(web): extract agent runtime contracts to break lib cycles
Create agent-contracts layer for ToolContext, repository task types,
write-gate policy, and GitHub text parsing helpers. Point agent-runtime
at contracts instead of agents/* and slim tools/types to re-exports.

HL-449
